+++ test/sources_test.exs
@@ -5,9 +5,9 @@ defmodule Sentry.SourcesTest do
 
   test "exception makes call to Sentry API" do
     correct_context = %{
-      "context_line" => "      raise RuntimeError, \"Error\"",
-      "post_context" => ["    end", "", "    post \"/error_route\" do"],
-      "pre_context" => ["", "    get \"/error_route\" do", "      _ = conn"]
+      "context_line" => "    raise RuntimeError, \"Error\"",
+      "post_context" => ["  end", "", "  post \"/error_route\" do"],
+      "pre_context" => ["", "  get \"/error_route\" do", "    _ = conn"]
     }
 
     bypass = Bypass.open()
@@ -25,7 +25,7 @@ defmodule Sentry.SourcesTest do
                |> Map.take(["context_line", "post_context", "pre_context"])
 
       assert body =~ "RuntimeError"
-      assert body =~ "DefaultConfig"
+      assert body =~ "Example"
       assert conn.request_path == "/api/1/store/"
       assert conn.method == "POST"
       Plug.Conn.resp(conn, 200, ~s<{"id": "340"}>)
@@ -35,7 +35,7 @@ defmodule Sentry.SourcesTest do
 
     assert_raise(Plug.Conn.WrapperError, "** (RuntimeError) Error", fn ->
       conn(:get, "/error_route")
-      |> Sentry.TestPlugApplications.DefaultConfig.call([])
+      |> Sentry.ExamplePlugApplication.call([])
     end)
   end
 end